Move text of CoL announcements to .yml
In an effort to release the announcement on the CoL hub as soon as possible,
the content was hardcoded in the template instead of being added to
the content_item.yml. This commit removes the hardcoded content, moves it
to content_item, rewrites the CoL announcement in the template and adds
'announcements' to the CostOfLivingLandingPagePresenter.