Add rails-i18n
Adds Rails I18n explicitly in the main section of the Gemfile.

Previously, it was a dependency on a few bundled Gems within the app but
was only available in developemnt / test.

Moving it has enabled us to use common translations in production, which
native Rails methods can take advantage of, e.g: `to_sentence(locale:
"ar")`.