Use blue bar from gem instead of pseudoelement
On Browse pages, the blue bar underneath the navbar was replicating the
style of an element from the govuk_publishing_components gem. This
resulted in some styling discrepancies between how the blue bar was
styled on browse pages versus the rest of gov.uk. This commit removes
the implementation of the blue bar in browse.scss and replaces it with
the blue bar from the gem directly.