Use footer link styles in announcement in CoL hub
Use the same class for the CoL announcement links as the footer links.
The contrast between the black links and grey background is higher
than that of the blue links and grey background. The footer uses the
same grey as the announcements, so I'm using the same classes for the
links.