Update the Plek service name for Search API
A [change] was name to gds-api-adapters to use the canonical name for
search-api, rather than "search".

Plek uses the service name in it's [environment variables] to find the
service. When the name of the service changed, Plek was unable to find
service and errored with:

```
GdsApi::SocketErrorException - Failed to open TCP connection to search-api.www.gov.uk:443 (getaddrinfo: nodename nor servname provided, or not known)
```

[change]: https://github.com/alphagov/gds-api-adapters/pull/1170
[environment variables]: https://github.com/alphagov/plek#environment-variables