Render logos for International Delegation pages
Where there are organisations linked to the World Location News content
item (i.e. for International Delegation pages), we need to render those
on the page.

This functionality has been migrated from Whitehall.  As part of this
change, we have fixed the crest issue (Whitehall incorrectly renders the
coat of arms for all departments, even if they have their own crest),
use the department's colour scheme and format the text as specified by
the department.