Use content_ids for curated lists
When a content item is tagged to a curated Mainstream browse or Specialist
topic and this content item’s slug changes, the link disappears from the topic
page. This is because the curated lists/groups are based on base paths. The
“contents” slugs/base paths don’t get updated as a part of dependency
resolution in Publishing API.

We want to use content_ids instead as they are less volatile.