Remove legacy css
The organisation__promotions css class is legacy styling that doesn't appear to still be in use.
So the parent column class logic that adds thoses classes to the template is not required. Git blame
does not shed light on what they were originally doing, but this is place in the code base those
styles are used, and removing them has no impact.