Update GitHub Action workflow for CI
This updates the initial workflow to enable CI using GitHub Actions. It
makes use of reusable workflows and composite actions defined in
govuk-infrastructure. The CI workflow is initially only used to support
deployments to the new EKS infrastructure and not to immediately replace
the CI workflow in Jenkins.

The CI workflow makes use of individual jobs to run each testing and
linting tool. This allows the processes to run concurrently and improves
the visibility of status in the GitHub UI. The workflow runs a precursor
job to cache dependencies in order to speed up other jobs.

Most jobs make use of reusable workflows as they are consistent across
the majority of our repositories - this allows us to easily maintain
them in a single place. However, the test-ruby job (and/or
integration-tests) is defined per repo as backing services and other
dependencies vary greatly between repos.  Backing services (such as
databases) are created using a composite action instead of a job
service. This is because they contain a lot of shared config and makes
it easier for us to maintain.