Simplify the show template for organisations
The show template was becoming hard to follow partly because of the length of the file and also because of a large number of conditional blocks of logic depending on whether the organisation is promotional not. We're currently working on adding a new organisation header to collections so this commit is to prep adding some more logic to organisations/show/.

Move the non-organisational content into its own template files. Remove a few now redundant conditional checks for whether an organisation is promotional or not
i