Add rack env for test environment
This runs slimmer in 'strict' mode, meaning that it will raise a runtime exception when there is invalid html (e.g. a div is not closed properly). This means that when you run 'bundle exec rspec' you should get the same result as running 'bundle exec rake spec' - without it this can diverge confusingly.