Change contact_form in en/organisations.yml
As part of the changes to the number 10 organisation page, the link to
the contact form on organisation pages has been changed to 'Contact:'
from 'Contact form:'. This is because sometimes the title of the contact
form would include 'form' already.

Updated spec test that was expecting the link to be 'Contact form:' to
check for 'Contact:' instead.