Demo of Contentful controlled Cost of Living Hub
This configures a rendered Cost of Living hub page based on the
Publishing API having been populated with data from Contentful.

It relies on a number of Govspeak fields to simplify both the input and
output of content. From running this there seems to be only a small
visual difference compared to the current page - which is that the
announcements links are shown in blue rather than black. I imagine there
may also be some functionality differences for analytics that may need
checking.

The Contentful model that was used to render this was:

Cost of Living Hub Page:

- title (short text)
- meta description (short text)
- lede (short text)
- announcements_govspeak (long text)
- accordions (orderable references to Cost of Living Accordions)
- footer_govspeak (long text)

Cost of Living Accordion:

- heading (short text)
- content_govspeak (long text)