Changes to featured documents on No. 10 org page
Remove the first_featured_document from No. 10 org pages. For No. 10 org
page there will be a hardcoded header so we don't need a large styled
image card with the first featured document.

Use small font size for image card component for the featured documents
on the No. 10 organisation page. This is also because of the upcoming
header change to the page.

Add tests for the change to documents_presenter.