Fix No10 header image card w/ image margins & width
The image card should work without an image but on mobile the margins aren't correctly set (there is extra margin added on mobile and the card has a contrained width, presumably to accomodate an image which shouldn't happen without an image).

We should investigate this bug separately and add a fix to the gem for it when we're confident that we understand all the use cases for the image card without image as the current behaviour could be serving a particular use case.