Make collections show default main language in analytics
Previously we would filter out the language from analytics if it
was our default `en` entry. While clean, it's inconsistent with
other areas of GovUK, so this has been changed to return the current
language regardless of whether it's the default.

The test was updated to ensure that `en` is the default locale, rather
than looking for a nil entry. This matches other behaviour, wherein we
rely on the default locale to be `en` rather than explicitly setting it.