Refactor historic appointments
There are three similar pages on gov.uk that show an index of past role holders of major offices of state: past prime ministers, past chancellors, and past foreign secretaries. At the moment, only past prime ministers is rendered from a full content item published by a publishing application - the other two pages are hard-coded in this application. The eventual end-state should be that none of these pages is hard-coded, and all are editable as a content item in a publishing application. The code for these three pages is therefore slightly more complex than if they were all rendered from a content item - the end state should be we have a single controller.

This commit refactors to combine code as much as possible, while attempting to not make the code confusing:

* moves past chancellors and past foreign secretaries into a shared controller, but keeps past prime ministers separate in order to not confuse code between hard-coded and non-hard-coded pages
* introduces a single index html template for use across all three pages, and associated presenter code
* moves some templates from shared into historic appointments, and rationalises down these templates for maximum re-use

The still share an index template, but the controllers are separate. When the hard-coding is removed, there should just be a single historic appointments controller that can render any of these pages.