Render /world/embassies
This commit adds rendering of /world/embassies, a page currently
rendered by Whitehall. When the embassies_index document is published
to the publishing API in production, we intend to switch over the
rendering of this route to this application (and remove the relevant
code from Whitehall).

We've designed the interface of EmbassiesIndexPresenter so that we can
reuse as much of the existing Whitehall view template[1] as
possible. Hopefully this makes the correspondence between the two as
clear as possible.

We have added tests for the two methods in EmbassiesIndexPresenter
that contain some logic, but we're relying on the higher level feature
spec to exercise the wrapping of json into objects in that
presenter. The fixture file we're using in the feature test is taken
from the example frontend embassies_index document in the publishing
api repo[2].

Although there are translations for the title of this embassies index
page in Whitehall (e.g. in en.yml[3]), there's no way to access any of
those translations because the /world/embassies route doesn't accept a
locale param[4]. Because of this we've chosen to hardcode the title of
the page in `EmbassiesIndexPresenter#title` to preserve the behaviour in
whitehall.

[1]: https://github.com/alphagov/whitehall/blob/a8a0634e6bf15596b43648e1f4cbcd302b73494d/app/views/embassies/index.html.erb
[2]: https://github.com/alphagov/publishing-api/blob/067cd760d76e91d97b47d5e8edecff2ac0fe5ea7/content_schemas/examples/embassies_index/frontend/embassies_index.json
[3]: https://github.com/alphagov/whitehall/blob/8ad388e7c9ad40283928df4baac0b5da8f172243/config/locales/en.yml#L446
[4]: https://github.com/alphagov/whitehall/blob/8ad388e7c9ad40283928df4baac0b5da8f172243/config/routes.rb#L48

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>
Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>