Remove the yellow COVID-19 top border temporarily
Yellow bar on COVID-19 landing page has been removed so that two bars
don't appear on the page when the global banner is deployed. There is
also a rule added for making sure there is not too much spacing between
the breadcrumb links and the navbar on mobile.

These changes are temporary and will be removed when the global bar has
been removed.