Practice setting up an AB test
- Displays different versions of the /coronavirus landing page depending on the test variant
- Please note - this will NOT be deployed to production

I tried to use the with_variant in the controller specs as per the docs [1] but encountered 'page is not defined, are you using capybara?' which happens when there is no page object [2]. Using with_variant in the controller tests did work when the controller test ran in isolation, but failed as part of bundle exec rspec. I spent a while trying to get capybara or active_support to work with the controller tests but gave up! I just set the request header manually instead.

[1] https://github.com/alphagov/govuk_ab_testing
[2] https://github.com/alphagov/govuk_ab_testing/blob/main/lib/govuk_ab_testing/acceptance_tests/capybara.rb#L8