Add routes for legacy organisation-related redirects
Trello: https://trello.com/c/sajxUZnZ

This will be used in conjunction with this whitehall PR [1] which
changes the main route for an organisation content item to be a "prefix"
route rather than a "exact" route. This change would mean that the
router would send the relevant URLs to collections instead of
whitehall-frontend. This PR adds the necessary routes which are closely
based on those currently in whitehall [2] so that collections can do the
redirects.

[1]: https://github.com/alphagov/whitehall/pull/7621
[2]: https://github.com/alphagov/whitehall/blob/9afe523ed2f3eca7771c5510056b4bbb2cfa29a1/config/routes.rb#L66-L73