Fix tests for profile groups in prime ministers index presenter
The implementation of these is correct (results in a page that is
rendered in the same way as the page previously was in whitehall) but
weren't caught due to a misnamed test file which resulted in the tests
not being run as part of the default rake task. In one case the shape of
the data we were expecting was updated in the implementation, but not in
the tests, and in another we were testing the wrong method.