Add i18n config for embassies index
This commit moves all hard-coded strings used to render the embassies
index view into the `en` locale file. Note that the `/world/embassies`
route doesn't take a `locale` parameter so it's impossible to render
the page using any of the other locales. However for consistency with
what's happening in the rest of the codebase, and to make the
`locales_validation_spec` pass we've added the corresponding empty
keys for all supported locales.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>