Show cabinet ministers
Show ministers' names and image cards.

Added styling to replicate the look of the `h2` from the original whitehall page.

Added `clear-both` style to get the columns within the grid to align as intended per the whitehall rendering. (Style also pinched from whitehall.)

In order to keep the commits readable, this commit doesn't include the
roles and payment info.