Ensure erb templates do not contain hard-coded strings
This uses Shopify's erb_lint gem[1] to ensure that the erb templates
in this project do not contain hard-coded, and therefore
untranslatable strings.

There's a lot of hard-coded content in the files under
historic_appointments/past_foreign_secretaries, and I suspect we don't
have any intention to offer translated versions of them. It would be a
bit painful to pull all of this content into localisation files so
I've decided to exclude them from the HardCodedString linter.

I investigated using rubocop-rails' (via `rubocop-govuk`) I18nLocalTexts
cop[2] for this, but unfortunately rubocop does not (by design) run
cops on erb templates.

[1] https://github.com/Shopify/erb-lint
[2] https://docs.rubocop.org/rubocop-rails/cops_rails.html#railsi18nlocaletexts