Show the lead paragraph on Ministers index
We are building upon the work started 3 years ago to bring rendering of
the Ministers index page from whitehall into collections.

We have already completed some additional work to add more data to the
content item[1], and now it had almost all the data we need to populate
this page. The only concept that is missing is the ordering of the whip
organisations within the Whips section.

I have discussed this with the team and we agreed that it's okay for
this knowledge to be encoded in the ministers index presenter.

[1] https://github.com/alphagov/publishing-api/pull/2359
https://github.com/alphagov/whitehall/pull/7624
https://github.com/alphagov/publishing-api/pull/2370