Show ministers' roles and payment type info
We are using an image card component, which makes the text slightly
bigger than the current whitehall styling. We think this is okay, on
account of the image card component being more modern and validated with
user research.

The info shown for role payment types replicates the way it is being
shown in whitehall.

We've also replicated the `joined_list` helper from whitehall, to get the
roles looking the same, i.e. using a semicolon as separator when any of
the roles contains a comma.