Don't show contacts section on international delegations page
We've received a request [1] from FCDO to remove the display of contacts from international delegation pages, since there's already a link to a separate page with contact details which appears just on it on the page. This removes a design that was ported over directly from Whitehall's rendering of the page.

1: https://govuk.zendesk.com/agent/tickets/4772368