Make organisation list filtering JavaScript more generic
Before this JavaScript was moved from Whitehall to Collections, it was shared
between the `/world` page and the `/government/organisations` page. However,
since being moved it has been made specific to the `/government/organisations`
page.

Now that we are also moving the `/world` page to Collections, we want to make
this module more generic so that it can be shared again.

This also removes `data-filter-list` as it is unused.