Disallow any domain from embeding a page to prevent clickjacking
This directive was initially added to enable side-by-side browser, which is a
tool used during transition, can work correctly. See:
https://github.com/alphagov/collections/commit/89284b09cb6364d6ed364ff519a2b6b920173254
The side-by-side browser is now retired and we want to prevent iframing of
GOV.UK as a defence against clickjacking.

The Content Security Policy is initialised
in `config/initializers/content_security_policy.rb` It configures
frame-ancestors to prevent any domain from framing the content in
https://github.com/alphagov/govuk_app_config/commit/e845dedfe2