Add Transition to list of API consumers in docs
While removing the reference to the (recently retired) transition-config
project from this list, I noticed that the (still very much active)
Transition app wasn't already included in it.

Disclaimer: I know nothing about this API and I've only found mention of
Whitehall as the source of Transition's organisations data, but I'm
trusting this document's explanation that the API used to be served by
Whitehall to mean that Transition is in fact a consumer of this.