Use medium_resolution_url and high_resolution_url instead of constructing image urls based on size
The way Whitehall references assets in Asset Manager - including images - is changing.
Previously Whitehall would have a hardcoded configuration for urls that would make it possible
for the various sizes of the same image urls to be guessable. The new behaviour is that each individual
image will have its own url provided by Asset Manager. The urls are therefore no longer guessable
as they all have a unique id in them.

The publishing-api schema allows for 3 fields to be sent: `url`, `medium_resolution_url` and `high_resolution_url`.
This commit ensures that the `url` is used where the original image asset is required, the `medium_resolution_url`
is used when the `s465` version is required and the `high_resolution_url` when the `s712` is required.
If additional sized are required, the `whitehall` presenters and `publishing-api` schemas would need to be updated.
For now, it seems like the 3 allowed fields satisfy the user needs.

Since `url` is the only required field in `publishing-api`, it is used as fallback when the other 2 are not available.

Co-authored-by: Iris Lau <iris.lau@digital.cabinet-office.gov.uk>
Co-authored-by: Tuomas Nylund <tuomas.nylund@digital.cabinet-office.gov.uk>