Add use_v1 to world_locations_news search links
The v2 search api isn't finding any documents for queries like

    https://www.gov.uk/search/all?order=updated-newest&world_locations%5B%5D=uk-delegation-to-council-of-europe

This is presumably because it doesn't have the same logic for indexing
world locations as v1 did. We should probably fix that and remove the
use_v1 flag, but for the short term this is a quick fix for the
immediate issue.

This was raised by FCDO on Zendesk[1]

- [1] https://govuk.zendesk.com/agent/tickets/5664229