Add guards to fix preview of Ministers page during a reshuffle
Currently, the rendering view assumes there's only 2 views of the page - a fully assigned ministers index, or a banner that is displayed in case of a reshuffle. In reshuffle mode preview, however, means that the if statement in the view assumes wrongly about the data, since the reshuffle flag / message is not passed for the preview. We should cater for these cases during reshuffle where there could be empty lists of ministers.