Prepare custom banners for specific open and closed organisations
There may be a requirement to add banners on a per-organisation
basis. This commit gets the groundwork in place for that, taking
inspiration from the last time we did this (in #3172). Unlike last
time, we may also need to have custom banners on _closed_
organisations, which already have their own banner logic.

Trello: https://trello.com/c/M1R5xg3w/2725-prepare-banners-for-new-and-closing-organisations