Retrospectively write unit tests for Ministers Index Presenter
There's a fair bit of logic encoded in this view/presenter, so
retrospectively adding some tests to ensure we don't break the
"default to empty array" fixes we've added recently to support
previewing.

Still need some tests for the 'Minister' embedded class, but that
begs wider questions such as "should this be moved out to its own
model file?" and so on. That can be tackled later.