Prepare locale file for popular task data
Popular tasks are currently present on business and benefits pages only.
As part of our AB test, we will be adding 3 popular tasks to all level 2
browse pages. This commit preps the file for the incoming data.

govuk-docker-run rake translation:add_missing has not been run - so the keys are
missing from other locale files. This is intentional. Browse pages are not
currently translated, and there is no intention of implementing translations. This
AB test code is temporary and will be removed in a few weeks. Adding empty keys
will just make the diff harder to review and reverting the code more time consuming.