Simplify control data code, and add tests
For variants A and B, all browse pages will contain popular links.
For variants C and Z ie the control state, only business and benefits will
contain popular links. The other 14 will default to hiding popular tasks
entirely - as this is the default behaviour currently live on GOV.UK

This commit adds specs to highlight this, and simplifies the existing code. There
is no behavioural change.