Render GOV.UK chat on some step-by-step nav pages
If the SBS nav page's `base_path` matches one in a predefined list, we
want to render the GOV.UK Chat promo to point users to the new chat
service. This adds the markup to the show template.

The designs call for the main header to be shown, then below that the
intro on the left and the promo on the right, hence why the markup had
to be split up a bit.

Previously it was a single 2/3 width `div` with the header above the
intro text.

Now we need to render a 2/3 width `div` with just the header, then below
that a separate row with a 2/3 width `div` for the intro and a 1/3 width
`div` containing the promo.