Remove any empty responses for a url from search
This code currently doesn't contain any guard clauses or handling of
error codes from SearchAPI which obviously wouyld be needed if we do this
for reals. This `compact` just removes any nils that have been
mapped to the links array following an exernal nhs page being sent to search api
for popularity data.

Presumably we won't support external links on browse pages
but as search looks up pages based on base_path, rather than content id, there is the
possibility that search returns a 404, so we'd need to think how to handle that.