Fetch data from BigQuery
- Create a popular tasks name space for these two services

- Ideally biqquery would have a simple table per browse page, and Collections would just be scraping the top 6 links.
Would need to chat to a PA to see what table structure is most cost effective (maybe better to have a single table for all browse)
For performance reasons, I think we should be relying on the processing of data at the big query side, and Collections should just
fetch the minimum amount of data with little/no processing at the application side.
This is a different approach to that taken in https://github.com/alphagov/collections/pull/3761/commits/046d6186287bd24b5366b515fedb0c0f131aa54e

This commit needs some work on the tests.