Use `first_featured_news` to get first featured item on No 10 page
Makes better use of the `first_featured_news` on the number 10 org page.
The `remaining_featured_news` method then does what it says and returns
all of the remaining news, rather than actually returning everything.

`large: true` is being merged into the first item in
`first_featured_news`, so that doesn't need to be repeated in the view.

I'm not sure if these changes are having an affect on how the page is
rendered.