Use GraphQL for rendering World Index page
This will allow us to build a proof of concept for using GraphQL on
GOV.UK.

It will run alongside the existing retrieval from Content Store to allow
us to switch the rendering used in different environment, through a
feature flag or A-B test.

The feature is switched on by adding a `GRAPHQL_FEATURE_FLAG`
environment variable.

Tests have been copied (as much as possible) from the existing tests
covering the Content Store version of the page, to ensure we are
retaining as much functionality as possible.

Note: the route has been moved further up the routes file, since
`DocumentTypeRoutingConstraint` requires the content item.