Fix faulty test expectation
This test assumes that `The Gambia` will always be the first item in the
`G` grouping.

However that may not be the case if the fixture data is updated to have
another country, e.g. Gabon.

Therefore updating the test to check whether this value appears anywhere
in the `G` section, not just first.

This matches the implementation in the GraphQL version of this spec.