Introduce a Graphql-specific Role model
The `content_item` attribute is consistent with the public interface of
the existing (Content Store-specific) Role model.

The `content_item` itself won't be the same type of object as
Role#content_item (which, FWIW, is a ContentItem), but it will allow us
to drop instances of this model into a few helper methods that would
otherwise require tricky code changes.

Co-authored-by: Ynda <yndajas@gmail.com>