Full width version of GOV.UK chat banner on browse pages
There is a desire to change the design of the GOV.UK Chat promo on just
the mainstream browse pages to be a full width one. This design is
intended to last only until 20th December when the banners will be
removed so this implementation priorities pace of delivery over
reusability.

To achieve this full width approach I have put the existing component
within a banner and done a couple of style overrides. This does have an
undesirable consequence that a change to the component will break this
app - however as this is expected to only be around for a matter of days
this risk seems tolerable.

In removing the other positioning of the banner I have gone to efforts
to restore the HTML to how it was before any chat banners were added.