Update GraphQL approach for world index
Given we're still in the learning stage with the GraphQL project, we've
decided to minimise the changes needed on frontend apps by retaining the
current heavily nested data structure. This updates Collections to work
with that format (the changes to Publishing API are also in progress).
This means that the code for the world index page in collections will be
less divergent between paths for GraphQL and REST data sources. This
commit undoes some of the divergence we initially introduced

We plan to revisit simplifying the data structure later (into a less
nested form)