Remove "important board members" image constraint
This feature was added in #754 (counterpart in Whitehall:
https://github.com/alphagov/whitehall/pull/4162). Neither PR
really explains/justifies the feature: it appears to have been
added as a workaround for a specific organisation, where they'd
neglected to add an image for one of their board members.
Instead of adding an image for them, this feature was added to
impose an artificial limit on the number of images to display on
the page.

This has since caused numerous support tickets for the publishing
team, with publishers struggling to understand why board member
photos are missing from the org page. The limit, tucked away in
the organisation settings, is not well known and is not intuitive.

We've decided to remove the feature, removing complexity from
both Whitehall and Collections. It should result in fewer support
tickets for the team, and be a net positive for organisations
in terms of an easier-to-understand system. If we find there is
a requirement to re-introduce something like this in future, we
can approach it more holistically. For now, the simplest thing is
to remove the feature, as we suspect it's not widely used.

Trello: https://trello.com/c/mz0WLkIR/3282-remove-display-management-team-images-feature