Add call to new template defined in govuk_web_banners
The recommended approach is to add a call to the partial to the application
layout below the breadcrumbs[1]

However on browse pages, the breadcrumbs sit inside the black browse header so adding the banner
across the header looks bad. For browse pages, we need to call the intervention banner partial below
the browse header.

[1] https://github.com/alphagov/govuk_web_banners?tab=readme-ov-file#adding-recruitment-banners