Attempt to refactor the roles in person model
Since we have added some methods to get previous non ministerial roles,
I have tried to refactor some of the methods to get the various roles of
a person.

I'm not sure if I've made it better or worse, so looking for comments on
the approach and any suggestions for improvement.