Populate meta tags on GraphQL version World Index
Until now, the GraphQL version of the World Index page has been
rendering with a few less meta tags than the regular version.

(I diff'd their respective HTML and looked up the missing meta tags in
GOV.UK Publishing Components to find the field names that needed adding
to the query.)

---

I get this test is reaching into the details of the components gem, but
it seems like the most obvious place right now to cover the interaction
of the fixtures, Collections logic and the component.

I can throw out the test if it's going to be too brittle.