Test that World Index renders meta_tags component
I get that this is overreaching into the details of the components gem,
but I think this is the only place that we can actually test this.

It covers the interaction of our example content item / GraphQL fixtures
with Collections and the meta_tags component.

I'm happy to throw out this commit if it doesn't sound like a good idea,
though.