Fix whip logic for GraphQL rendered Ministers Index page
For GraphQL responses, we are currently incorrectly showing non-whip
roles for people in the whips section of the ministers index page.

This is because GraphQL provides a hash for the `whip_organisation`,
with each value being `nil`. We cannot change this, as that is a
fundamental GraphQL concept.

Therefore updating the logic to check for a label, instead of just the
presence of an organisation hash.