Allow `ordered_parent_organisations` links to be an empty array
The code is currently assuming the `ordered_parent_organisations` links
will be nil if there are no parent organisations.

This assumption is not correct, as the link could theoretically be
present but with an empty array.

Therefore updating to support both cases.

This is important for GraphQL, since we will always be including an
empty array, even if there are no links.