Refactor ministers index department specific roles
This commit is a refactoring (i.e. there should be no changes to behaviour).

The ministers index page shows each minister's roles multiple times. Firstly, when the minister is listed in "cabinet ministers" (or whichever heading), it lists every current role the minister has. Secondly, when the minister is listed under the department they belong to it lists only the roles related to that department.

Previously we had an "org_role_ids" field on the minister object, and the code would construct different minister objects depending on whether they were going to be displayed on their own (in which case org_role_ids should be nil), or as part of a department (in which case org_role_ids should be all the content_ids for roles within the department).

I think this is kind of confusing object modelling - a minister should just be a minister, and they have the roles they have. We should worry about which roles to display when we come to display them, not when we build the minister object.

This commit adds a roles_for_department(department) method, which does the filtering, and we call this from the helper that renders minister descriptions. We only provide the department in the partial that is used to render department specific roles, so the decision on whether to use the filter is much closer to the view than before.

Hopefully this is slightly easier to understand than before, which will support some graphql-specific improvements which I would like to make down the line.