Add a more efficient way to select roles by department
Currently, content-store returns every role in each department, and we join this on the roles each minister has to work out the ministerial roles relevant to the department.

This is highly inefficient, because the vast majority of the roles in a department are board members and other functionaries who don't appear on the ministers index page. So we're fetching a huge amount of data we don't need.

With the graphql implementation, we can follow the link from each role back to the organisations it belongs to. So instead of saying "of all the roles belonging to dept X, is this role one of them?" we can say "of all the departments this role belongs to, is dept X one of them?".

Unfortunately, the link expansion that populates content-store doesn't allow us to do this. It won't expand cycles of links, so you can't go: department -> roles -> departments because it will ignore the current department due to the cycle.

This commit supports both patterns, so if a list of org_role_ids is present in the response it will use that, and if not it will try to use the org_ids for the role.

This allows us to take advantage of graphql's ability to fetch the links we want, while still supporting content-store.