Include taxonomy meta tags for GraphQL roles
Some role pages include meta tags representing the taxonomy that the
role belongs to.

We therefore need to request the `taxons` links as part of the GraphQL
query to retrieve the content for these meta tags.

Adding five levels of taxonomy, as was done in Frontend [1], since this
is the largest level of taxonomy we currently have.

1: https://github.com/alphagov/frontend/pull/4741/commits/68b98ece36c05258f06fddef7c23e14ce97089f3