Remove cookies from GraphQL A/B test
We are currently using the GOV.UK A/B testing framework to run an A/B
test for GraphQL.

This relies of users opting into analytics cookies, so we are seeing
very little traffic. Especially as these pages represent very small
levels of traffic in the first place.

We don't actually need cookies for this A/B test, since the changes are
all behind the scenes, so users won't see any difference.

Therefore updating the GraphQL A/B testing code to assign a variant for
each request, without needing any cookies.

The initial traffic rates have been selected by using the rate of
traffic seen to each document type's GraphQL variant in the last 3 days.
This will be adjusted once the new A/B test is in place.