Add test for new past prime minister sorting
Add a test for the past prime ministers presenter to check that prime
ministers with the same start year are correctly sorted.

This test does not use the `api_data_for_person_without_historical_account`
method like the other tests in the file since it does not have
functionality to add an end_year.